fix: show exact recipe purchase quantities

+++ apps/agent/src/domain/schemas.ts
@@ -59,7 +59,8 @@ export const partyPlanningInputSchema = z.object({
 
 export const plannerSelectionSchema = z.object({
   productId: z.string().min(1),
-  quantity: z.number().positive(),
+  // Count of catalog purchase increments/packages, never kilograms or a free-form recipe amount.
+  quantity: z.number().int().positive(),
   assignedMemberIds: z.array(z.string().min(1)).min(1),
   reason: z.string().min(1),
 });
